Add files via upload
CandyDatabase- 
The app will have the ability to add, delete, and update products in real time and save them into a database. 

Tiger King App- 
This one is just pretty funny. I made a soundboard from some main characters from Netflix Tiger King show. It plays on loop of them talking and you can pause and play others. 

Shipping Calc App-
This app calculates the cost for an item to be shipped. You enter the weight and it gives you the cost to ship it. ( more weight more cost) 

BBQ Recipes-
Since I am an avid BBQ smoker, I built this app to click on a food item and it will tell you the recipe. The pictures are really my food I have made on my BBQ smoker.

Number Calc-
This app can do 3 cool options. It can find the larger of 2 numbers, check if the number is divisible by 2, and it will calculate the factorial of the number and display all the answers.